refactor: make Index system ITensors.jl-compatible (u64 ID, id+tags equality)
Align the index system with ITensors.jl (non-QN version) for HDF5 format
compatibility. Three key changes:

1. DynId: u128 → u64 (matches ITensors.jl's IDType = UInt64)
2. Equality/Hash: id-only → id+tags (matches ITensors.jl's id+plev+tags
   with plev conceptually fixed at 0)
3. Hash implementations added to SmallString and TagSet

Updates across all layers: core Rust, C API, Julia bindings, Python
bindings, tests, and documentation.

+++ crates/tensor4all-core/src/defaults/tensor_data.rs
@@ -332,7 +332,7 @@ mod tests {
     fn new_id() -> DynId {
         use std::sync::atomic::{AtomicU64, Ordering};
         static COUNTER: AtomicU64 = AtomicU64::new(1);
-        DynId(COUNTER.fetch_add(1, Ordering::Relaxed) as u128)
+        DynId(COUNTER.fetch_add(1, Ordering::Relaxed))
     }
 
     #[test]
